临时提交，删除了T的泛型定义

+++ metric-calculate-core2/src/main/java/com/yanggu/metric_calculate/core2/enums/BasicType.java
@@ -31,7 +31,7 @@ public enum BasicType {
     /**
      * 小数型
      */
-    DECIMAL(BigDecimal.class),
+    DOUBLE(Double.class),
 
     /**
      * 字符串集合
